\name{Internal Functions for pbdPROF}
\alias{get.conf}
\alias{get.lib}
\title{ All Internal Functions for pbdPROF}
\description{ All internal functions for pbdPROF}
\references{
  Programming with Big Data in R Website:
  \url{http://r-pbd.org/}
}
\author{
  Wei-Chen Chen \email{wccsnow@gmail.com}.
}
\keyword{internal}

